% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageQueue.R
\name{appendMessage.MessageQueue}
\alias{appendMessage.MessageQueue}
\title{Title appendMessage.MessageQueue}
\usage{
\method{appendMessage}{MessageQueue}(x, message = NA, level = NA)
}
\arguments{
\item{x}{A MessageQueue object}

\item{message}{A message or vector of messages.}

\item{level}{A string or vector of message levels (ERROR, WARNING, INFO)}
}
\value{
A MessageQueue object
}
\description{
Internal S3 method to deal with appending messages
}
