% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageQueue.R
\name{appendMessage}
\alias{appendMessage}
\title{Title appendMessage}
\usage{
appendMessage(x, message, level)
}
\arguments{
\item{x}{A message queue}

\item{message}{A string or vector of strings of messages.}

\item{level}{A string or vector of strings of
message levels (ERROR, WARNING, INFO)}
}
\value{
A MessageQueue class.
}
\description{
Generic function to handle appending messages to a MessageQueue
}
