% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementCadence.R
\name{checkDataElementCadence}
\alias{checkDataElementCadence}
\title{Check Data Element Cadence}
\usage{
checkDataElementCadence(
  d,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{D2 Parsed data frame}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame of "dataElement","period",
of invalid data elements which are present the the data, if any.
If there are no violations, a boolean TRUE is returned.
}
\description{
Utility function to produce a data frame of
invalid data value based on the so-called data element cadence.
Quarterly data elements can be submitted each quarter.
SAPR data elements should only be submiteted in FYQ2 and FYQ4.
APR data elements should only be submitted in FYQ4.
}
\examples{
\dontrun{
  d <- d2Parser("myfile.csv",type="csv")
  checkDataElementCadence(d)
}
}
