% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementMechanismValidity.R
\name{checkDataElementMechValidity}
\alias{checkDataElementMechValidity}
\title{checkDataElementMechValidity(data,organisationUnit,datasets, d2session)}
\usage{
checkDataElementMechValidity(d, datasets, d2session = d2_default_session)
}
\arguments{
\item{d}{datimvalidation object}

\item{datasets}{Should be a character vector of data set UIDs.}

\item{d2session}{datimutils d2session object}
}
\value{
datimvalidation object
}
\description{
Returns a data frame invalid data elements which exist in the data
but which do not have a valid association with an attribute option combo (mechanism.)
}
\examples{
\dontrun{
     d <- d2Parser("myfile.csv", type = "csv")
     ds <- getCurrentDataSets(type = "RESULTS")
     d <- checkDataElementOrgunitValidity(data = d,datasets = ds)
}

}
