% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{checkDataElementOrgunitValidity}
\alias{checkDataElementOrgunitValidity}
\title{Check Data Element Organisation Unit Validity}
\usage{
checkDataElementOrgunitValidity(
  data = NA,
  organisationUnit = NA,
  datasets = NA,
  return_violations = TRUE,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{data}{D2 compliant data frame object}

\item{organisationUnit}{Organisation unit ancestor.}

\item{datasets}{Should be a character vector of data set UIDs.
Alternatively, if left missing, user will be promted to choose from a list.}

\item{return_violations}{Boolean to return violations only.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns subset of data which contains
  invalid data element / organisation unit associations. If no violations are
  found, a boolean TRUE value is returned.
}
\description{
Returns a data frame invalid data elements which exist in the
data but which do not have a valid organistion unit / dataset association.
}
\examples{
\dontrun{
     d <- d2Parser("myfile.csv",type="csv")
     ds <- getCurrentMERDataSets(type="RESULTS")
     checkDataElementOrgunitValidity(data=d,datasets=ds)
}
}
