% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMechanismPeriodValidity.R
\name{checkMechanismPeriodValidity}
\alias{checkMechanismPeriodValidity}
\title{checkMechanismPeriodValidity(data)}
\usage{
checkMechanismPeriodValidity(data)
}
\arguments{
\item{data}{A data frame which has been parsed by either d2Parser or sims2Parser}
}
\value{
Returns a data frame of invalid period-mechanism combinations.
}
\description{
This function will return an object of invalid mechanisms and periods. All data which is reported 
must have a period within the valid start and end dates of the attribute option combination to which it is assigned.
}
