% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNarrativeLength.R
\name{checkNarrativeLength}
\alias{checkNarrativeLength}
\title{Title Check that the length of narrative data does not exceed 50k characters.}
\usage{
checkNarrativeLength(
  data,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}
}
\description{
Title Check that the length of narrative data does not exceed 50k characters.
}
