% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOperatingUnit.R
\name{checkOperatingUnit}
\alias{checkOperatingUnit}
\title{Function which checks that the organisation unit provided is valid}
\usage{
checkOperatingUnit(
  organisationUnit = NA,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{organisationUnit}{Organisation unit ancestor.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a vector of organisation unit UIDs
}
\description{
Validates that the supplied UID is a valid operating unit UID
}
\note{
remapOUs(foo,"https://www.datim.org","admin",
         "district","code","id","Ab12345678")
will remap organisation units specified as codes to UIDs
}
