% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOrgunitsInHierarchy.R
\name{checkOrgunitsInHierarchy}
\alias{checkOrgunitsInHierarchy}
\title{Title}
\usage{
checkOrgunitsInHierarchy(
  d,
  userOrgUnit,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}
}
\description{
Title
}
