% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodUtils.R
\name{checkPeriodIdentifiers}
\alias{checkPeriodIdentifiers}
\title{Check Period Identifiers}
\usage{
checkPeriodIdentifiers(d)
}
\arguments{
\item{d}{D2 Parsed data frame}
}
\value{
TRUE if all periods are valid.
}
\description{
Expect an error if any invalid period identifiers are
supplied in the file.
}
\examples{
\dontrun{
    d <- d2Parser("myfile.csv", type = "csv")
    #Should return no error if all period identifiers are valid
    checkPeriodIdentifiers(d)
}

}
