% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidationRules.R
\name{checkValidationRules}
\alias{checkValidationRules}
\title{Title}
\usage{
checkValidationRules(
  d,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{datimvalidation object}

\item{d2session}{datimutils login object}
}
\value{
datimvalidation object
}
\description{
Title
}
