% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{checkValueTypeCompliance}
\alias{checkValueTypeCompliance}
\title{checkValueTypeCompliancep(d,base.url,username,passowrd)}
\usage{
checkValueTypeCompliance(d, secrets)
}
\arguments{
\item{d}{Parsed data frame}

\item{secrets}{Location of DHIS secrets file}
}
\value{
Returns a data frame of invalid data only
}
\description{
Utility function of extraction of data element ids, codes, shortName and names
}

