% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{clearCache}
\alias{clearCache}
\title{clearCache(wd)}
\usage{
clearCache(cache_dir = paste0(getwd(), "/.R_Cache/"))
}
\arguments{
\item{wd}{Cache directory path. By default, working directory + ".R_Cache"}
}
\value{
Nothing.
}
\description{
Internal utility function to clear the object cache.
}

