% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2_api_get.R
\name{d2_api_get}
\alias{d2_api_get}
\title{Execute and return a DATIM API query.}
\usage{
d2_api_get(path, d2session, timeout = 60)
}
\arguments{
\item{path}{Should exclude the "api" portion of the HTTP request. This will
be added as part of the GET request.}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logining in to datim with loginToDATIM}
}
\value{
Result the result of the API query as a list.

Simplified version of the original api_get function from datimutils.
}
\description{
Execute and return a DATIM API query.
}
