% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{datim_validation_params}
\alias{datim_validation_params}
\title{Standardized package function parameter definitions}
\usage{
datim_validation_params(
  csv_header,
  d,
  data,
  dataElementIdScheme,
  datasets,
  filename,
  hasHeader,
  idScheme,
  invalidData,
  iso,
  ISO,
  isoPeriod,
  cocs_in,
  des_in,
  mechs_in,
  ous_in,
  mode_in,
  mode_out,
  organisationUnit,
  orgUnitIdScheme,
  return_violations,
  return_violations_only,
  sig,
  vr,
  wd,
  d2session,
  d2_session
)
}
\arguments{
\item{csv_header}{By default, CSV files are assumed to have a header,
otherwise FALSE will allow for files without a CSV header.}

\item{d}{D2 Parsed data frame}

\item{data}{D2 compliant data frame object}

\item{dataElementIdScheme}{Should be one of either code, name, shortName,
or id. If this paramater is "id", then the Data elements are assumed to be
already specififed as UIDs.}

\item{datasets}{Should be a character vector of data set UIDs. 
Alternatively, if left missing, user will be promted to choose from a list.}

\item{filename}{Location of the payload to be imported. Should be a valid
import file.}

\item{hasHeader}{TRUE by default. Should be set to FALSE if the file does not
contain header information.}

\item{idScheme}{Remapping scheme for category option combos.}

\item{invalidData}{Exclude any (NA or missing) data from the parsed file?}

\item{iso}{String which identifies the period, such as 2016Q1 or 2016Q2}

\item{ISO}{Formatted string such as 2016Q1. Refer to DHIS2 documentation for
details.}

\item{isoPeriod}{period to be used for date shift boundaries. If not
provided, no boundaries are set.}

\item{cocs_in}{A vector of category option combinations.}

\item{des_in}{A vector of data element identifiers (codes, names, or
shortNames).}

\item{mechs_in}{A vector of data element identifiers (codes, names, or
shortNames).}

\item{ous_in}{A vector of organisation unit identifiers (codes, names, or
shortNames).}

\item{mode_in}{Should be one of code, name,shortName or id. This is the class
we are mapping from.}

\item{mode_out}{Should be one of code,name,shortName or id. This is the class
we are mapping to.}

\item{organisationUnit}{Organisation unit ancestor.}

\item{orgUnitIdScheme}{Should be one of either code, name, shortName or id.
If this paramater is "id", then the organisation units are assumed to be
already specififed as UIDs}

\item{return_violations}{Boolean to return violations only.}

\item{return_violations_only}{Parameter to return only violations or all
validation rule evaluations.}

\item{sig}{MD5 hash of the objects URL reference}

\item{vr}{Validation rule object}

\item{wd}{By default, the .R_Cache directory in the working directory.}

\item{d2session}{DHIS2 Session id}

\item{d2_session}{DHIS2 Session id}
}
\value{
list of all paramaters of this constructor function
}
\description{
Standardized package function parameter definitions
}
