% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datimvalidation.R
\docType{package}
\name{datimvalidation}
\alias{datimvalidation}
\title{DATIM Validation}
\description{
The datimvalidation package can be used to provide
external validation of data payloads intended for PEPFAR's instance
of DHIS2, located at https://www.datim.org.
}
