% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHISLogin.R
\name{DHISLogin}
\alias{DHISLogin}
\title{R6 Class which handles logging into to a DHIS instance.}
\description{
R6 Class which handles logging into to a DHIS instance.

R6 Class which handles logging into to a DHIS instance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config_path}}{Path to a JSON configuration file.}

\item{\code{baseurl}}{The URL of the server, e.g. https://www.datim.org/.}

\item{\code{username}}{Your user name.}

\item{\code{password}}{Your password.}

\item{\code{user_orgunit}}{UID of the users assigned organisation unit}

\item{\code{handle}}{An httr handle used to communicate with the DHIS2 instance.}

\item{\code{maxCacheAge}}{By default, set to 7 days. Used to control the valid 
age of objects which are cached locally.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DHISLogin$new()}}
\item \href{#method-login}{\code{DHISLogin$login()}}
\item \href{#method-clone}{\code{DHISLogin$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DHISLogin object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DHISLogin$new(config_path = NA, baseurl = NA, username = NA, password = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_path}}{Configuration file path}

\item{\code{baseurl}}{URL to the server.}

\item{\code{username}}{Your username}

\item{\code{password}}{Your password.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
Login to a DHIS2 instance and set the httr handle for the specified server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DHISLogin$login()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DHISLogin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
