% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateValidation.R
\name{evaluateValidation}
\alias{evaluateValidation}
\title{Utility function for evaluating validation rules}
\usage{
evaluateValidation(combis, values, vr, return_violations_only = TRUE)
}
\arguments{
\item{combis}{Data elements and category option combinations}

\item{values}{Values}

\item{vr}{Validation rule object}

\item{return_violations_only}{Parameter to return only violations or all
validation rule evaluations.}
}
\value{
Returns a data frame of validation rule evaluations
}
\description{
evaluationValidation will accept a vector of combis (data elements.category optioncombinations) and values
and evaluate it against the supplied validation rules
}
