% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSecrets.R
\name{GetCredentialsFromConsole}
\alias{GetCredentialsFromConsole}
\title{GetCredentialsFromConsole()}
\usage{
GetCredentialsFromConsole()
}
\value{
A list of baseurl, username and password
}
\description{
Obtains the credentials information from the console.
}
