% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{getCredentialsFromKeyring}
\alias{getCredentialsFromKeyring}
\title{getCredentialsFromKeyring(ring, service, username)}
\arguments{
\item{ring}{keyring name}
}
\value{
a list containing entries called password, baseurl, and username
}
\description{
retrieves username, service, and password from keyring
}
