% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrossWalkDuplicates}
\alias{getCrossWalkDuplicates}
\title{getCrossWalkDuplicates(d,base.url,username,password)}
\usage{
getCrossWalkDuplicates(d, base.url, username, password)
}
\arguments{
\item{d}{A d2 parsed data frame of data.}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a data frame of all crosswalk mechanism values, along with default de-deudplication adjustments.
}
\description{
Utility function which returns the crosswalk mechanism UID
}

