% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMap}
\alias{getCrosswalkMap}
\title{getCrosswalkMap(base.url,username,password)}
\usage{
getCrosswalkMap(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a map of DSD/TA data element UIDs.
}
\description{
Utility function which returns a map of DSD / TA data elements UIDs
}

