% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMechanism}
\alias{getCrosswalkMechanism}
\title{getCrosswalkMechanism()}
\usage{
getCrosswalkMechanism()
}
\value{
Returns a UID of the crosswalk de-duplication mechanism
}
\description{
Utility function which returns the crosswalk mechanism UID
}

