% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getCurrentDataSets}
\alias{getCurrentDataSets}
\title{Get Current datasets provided a data stream}
\usage{
getCurrentDataSets(
  datastream = "RESULTS",
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{datastream}{One of "RESULTS", "TARGETS", "NARRATIVES", "SIMS".
RESULTS and TARGETS refer to the current MER Results and targets datasets
NARRATIVES refer to the current MER Results Narratives
SIMS refer to the current SIMS datasets.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a list of dataset UIDs of the given type
}
\description{
Get Current datasets provided a data stream
}
\examples{
  \dontrun{
 ds <- getCurrentDataSets(type="RESULTS", d2session = d2session)
 }

}
