% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getCurrentMERDataSets}
\alias{getCurrentMERDataSets}
\title{getCurrentMERDataSets(type)}
\usage{
getCurrentMERDataSets(type = "RESULTS", creds)
}
\arguments{
\item{type}{Should be either RESULTS or TARGETS}

\item{creds}{DHISLogin object.}
}
\value{
Returns a list of dataset UIDs of the given type
}
\description{
getCurrentMERDataSets(type)
}
\examples{
  \dontrun{
 ds<-getCurrentMERDataSets(type="RESULTS", creds = creds)
 }
 
}
