% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataElementMap.R
\name{getDataElementMap}
\alias{getDataElementMap}
\title{getDataElementMap(base.url,username,passowrd)}
\usage{
getDataElementMap(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a data frame  of id,code,shortName and name
}
\description{
Utility function of extraction of data element ids, codes, shortName and names
}

