% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{getDataElementOrgunitMap}
\alias{getDataElementOrgunitMap}
\title{Title}
\usage{
getDataElementOrgunitMap(
  dataset,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}

\item{datasets}{Vector of dataset UIDs}
}
\value{
A list of data element IDs and all organisation units which are valid.
}
\description{
Title
}
