% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidDatasetMembers.R
\name{getDataElementsOrgunits}
\alias{getDataElementsOrgunits}
\title{getDataElementsOrgunits(data,base.url,username,password,organisationUnit,datasets)}
\usage{
getDataElementsOrgunits(base.url, username, password, organisationUnit,
  datasets = NA)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{organisationUnit}{Should be the UID of the organisation unit ancestor.}

\item{datasets}{Should be a character vector of data set UIDs. Alternatively, if left missing, user will be promted.}
}
\value{
Returns a data frame of "data" which was submitted along for a reason of why the data is considered to be invalid.
}
\description{
Returns a map of lists consisting of data elements and orgunits for a dataset (or datasets) for a given organisationUnit
}

