% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getDataSets}
\alias{getDataSets}
\title{getAvailableDataSets()}
\usage{
getDataSets()
}
\value{
Returns a data frame  of
}
\description{
Utility function to produce a data frame of valid data elements based on current
DATIM form specification
}

