% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementMechanismValidity.R
\name{getDE_ACOC_Map}
\alias{getDE_ACOC_Map}
\title{Title Get a list of data elemement and attribute option combos
which they are theoretically associated with from the server.}
\usage{
getDE_ACOC_Map(
  dataset,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}

\item{datasets}{Vector of dataset UIDs}
}
\value{
A list of all data elements and valid attribute option combos
based on a given dataset.
}
\description{
Title Get a list of data elemement and attribute option combos
which they are theoretically associated with from the server.
}
