% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getExactDuplicates}
\alias{getExactDuplicates}
\title{getExactDuplicates(data)}
\usage{
getExactDuplicates(d)
}
\arguments{
\item{d}{A d2 parsed data frame}
}
\value{
Returns a data frame of d2 data, with exact duplicates.
}
\description{
Returns a listing of exact duplicates by dataElement,period,orgunit,categoryOptionCombo and attributeOptionCombo
}
