% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidDatasetMembers.R
\name{getInvalidDataElementFromDatasets}
\alias{getInvalidDataElementFromDatasets}
\title{getInvalidOrgunitsFromDatasets(data,organisationUnit,datasets)}
\usage{
getInvalidDataElementFromDatasets(data, dataElement, datasets)
}
\arguments{
\item{data}{D2 compliant data frame}

\item{dataElement}{Should be the UID of the organisation unit ancestor.}

\item{datasets}{Should be a character vector of data set UIDs. Alternatively, if left missing, user will be promted.}
}
\value{
Returns a data frame of "data" which was submitted along for a reason of why the 
data is considered to be invalid.
}
\description{
Returns a data frame invalid data elements which exist in the data 
but not the datasets
}

