% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidDatasetMembers.R
\name{getInvalidOrgunitsFromDatasets}
\alias{getInvalidOrgunitsFromDatasets}
\title{getInvalidOrgunitsFromDatasets(data,organisationUnit,datasets)}
\usage{
getInvalidOrgunitsFromDatasets(data, organisationUnit = NA, datasets = NA)
}
\arguments{
\item{data}{D2 compliant data frame}

\item{organisationUnit}{Organisation unit. Defaults to user organisation unit if not supplied explicitly.}

\item{datasets}{Should be a character vector of data set UIDs. Alternatively, if left missing, user will be promted.}
}
\value{
Returns a data frame of "data" which was submitted along for a reason
 of why the data is considered to be invalid.
}
\description{
Returns a subset of data which exists in the data, but
which do not have corresponding organiation unit assignments
}

