% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechanismsMap.R
\name{getMechanismsMap}
\alias{getMechanismsMap}
\title{getMechanismsMap()}
\usage{
getMechanismsMap(organisationUnit = NA, ISO = NA)
}
\arguments{
\item{organisationUnit}{Operating unit in question specified as a UID. If not specified, assume to be Global.}

\item{ISO}{String which identifies the period, such as 2016Q1 or 2016Q2}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
If the ISO period is supplied, isValid will indicate whether the mechanism is valid for the given period
}
\description{
Utility function to produce a map of valid mechanisms
}

