% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOrganisationUnitMap.R
\name{getOrganisationUnitMap}
\alias{getOrganisationUnitMap}
\title{getOrganisationUnitMap(base.url,username,passowrd,operatingUnit)}
\usage{
getOrganisationUnitMap(base.url, username, password, organisationUnit)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{organisationUnit}{of question specified as a UID}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}

