% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOrganisationUnitMap.R
\name{getOrgansationUnitMap}
\alias{getOrgansationUnitMap}
\title{getOrgansationUnitMap(base.url,username,passowrd,operatingUnit)}
\usage{
getOrgansationUnitMap(base.url, username, password, operatingUnit)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{operatingUnit}{of question specified as a UID}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}

