% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodUtils.R
\name{getPeriodFromISO}
\alias{getPeriodFromISO}
\title{getPeriodFromISO(iso)}
\usage{
getPeriodFromISO(iso)
}
\arguments{
\item{ISO}{String which identifies the period, such as 2016Q1 or 2016Q2}
}
\value{
Returns a data frame consisting of iso (character),startDate (Date),endDate (Date) and period type (character)
}
\description{
Utility function return a data frame of a periods start date, end date, ISO character and periodtype
}

