% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodUtils.R
\name{getPeriodType}
\alias{getPeriodType}
\title{getPeriodType(iso)}
\usage{
getPeriodType(iso)
}
\arguments{
\item{iso}{String which identifies the period, such as 2016Q1 or 2016Q2}
}
\value{
Returns a a character of the period, like "Monthly"
}
\description{
Utility function return a DHIS2 period type from an ISO string
}
\examples{
getPeriodType("2018Q3")
getPeriodType("2018")
getPeriodType("19730221")
getPeriodType("201904")
getPeriodType("2018Q4")

}
