% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidDataElements.R
\name{getValidDataElements}
\alias{getValidDataElements}
\title{getInvalidDataElements(data,base.url,username,password,dataset)}
\usage{
getValidDataElements(base.url, username, password, dataset = NA)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{dataset}{Should be the UID of the dataset which you are validating.}
}
\value{
Returns a data frame  of "dataSet","dataElementName","shortname","code","dataelementuid","categoryOptionComboName"
}
\description{
Utility function to produce a data frame of valid data elements based on current
DATIM form specification
}

