% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{isValidCachedObject}
\alias{isValidCachedObject}
\title{isValidCachedObject(sig,wd)}
\usage{
isValidCachedObject(x)
}
\arguments{
\item{x}{file path of the cached object}
}
\value{
If the file exists and is not stale, returns TRUE, otherwise FALSE.
}
\description{
Internal utility function to determine whether a cached object is stale.
If its stale, it will be removed.
}
