% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{loadConfigFile}
\alias{loadConfigFile}
\title{LoadConfig(config_path)}
\arguments{
\item{config_path}{Path to the DHIS2 credentials file}
}
\value{
A parsed list of the configuration file.
}
\description{
Loads a JSON configuration file to access a DHIS2 instance
}
