% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExchangeFile.R
\name{loadExchangeFile}
\alias{loadExchangeFile}
\title{Title}
\usage{
loadExchangeFile(
  filename,
  type,
  datastream,
  organisationUnit = NA,
  dataElementIdScheme = "code",
  orgUnitIdScheme = "id",
  idScheme = "id",
  invalidData = FALSE,
  hasHeader = TRUE,
  isoPeriod = NA,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}
}
\description{
Title
}
