% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSecrets.R
\name{loadSecrets}
\alias{loadSecrets}
\title{loadSecrets(secrets)}
\usage{
loadSecrets(secrets = NA)
}
\arguments{
\item{secrets}{Location of the secrets file}
}
\value{
Returns a boolean value indicating that the secrets file is valid by accessing /api/me
}
\description{
Load a DHIS secrets file which should be a JSON file as below
{
dhis": {
 "baseurl": "http://localhost:8080",
 "username": "admin",
 "password": "district"
 }
}
}
