% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageQueue.R
\name{MessageQueue}
\alias{MessageQueue}
\title{Title MessageQueue}
\usage{
MessageQueue(message = character(), level = character())
}
\arguments{
\item{message}{One or more character strings}

\item{level}{One of ERROR, WARNING, INFO}
}
\value{
Object of class data.frame and Message queue.
}
\description{
A simple S3 object to deal with messages created during
DataPack processing
}
