% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageQueue.R
\name{printMessages.MessageQueue}
\alias{printMessages.MessageQueue}
\title{Title printMessage.MessageQueue}
\usage{
\method{printMessages}{MessageQueue}(x)
}
\arguments{
\item{x}{A MessageQueue object}
}
\value{
Returns a formatted output to the console
}
\description{
Internal S3 method to deal with printing messages
}
