% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageQueue.R
\name{printMessages}
\alias{printMessages}
\title{Title printMessages}
\usage{
printMessages(x)
}
\arguments{
\item{x}{A MessageQueue object}
}
\value{
Returns a formatted output to the console
}
\description{
Generic function to handle printing messages of a MessageQueue
}
