% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapDEs.R
\name{remapDEs}
\alias{remapDEs}
\title{Function which converts data element codes, names or shortnames to another class of identifiers}
\usage{
remapDEs(
  des_in,
  mode_in = "code",
  mode_out = "id",
  d2session = d2_default_session
)
}
\arguments{
\item{des_in}{A vector of data element identifiers (codes, names, or
shortNames).}

\item{mode_in}{Should be one of code, name,shortName or id. This is the class
we are mapping from.}

\item{mode_out}{Should be one of code,name,shortName or id. This is the class
we are mapping to.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a vector of DATIM uids for each data element which needs to be remapped.
}
\description{
remapDes should be supplied a vector of data elements (names,codes or shortnames)
along with the other required paramaters. It will return a vector of identifiers of a different class.
}
\examples{
\dontrun{
    d<-d2Parser("myfile.csv",type="csv")
    d$de_codes<-remapOUs(d$dataElements,mode_in="id",mode_out="code")
}
}
