% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapMechs.R
\name{remapMechs}
\alias{remapMechs}
\title{Function which converts mechanism codes, names or shortnames to UIDs}
\usage{
remapMechs(
  mechs_in,
  organisationUnit,
  mode_in = "code",
  mode_out = "id",
  creds
)
}
\arguments{
\item{mechs_in}{A vector of data element identifiers (codes, names or shortNames)}

\item{organisationUnit}{The UID of the operating unit.}

\item{mode_in}{Should be one of code, name or id. This is the type we are mapping from.}

\item{mode_out}{Should be one of code,name or id. This is the type we are mapping to.}

\item{creds}{DHIS Login credentials object}
}
\value{
Returns a vector of mechanism UIDs
}
\description{
remapMechs should be supplied a vector of data mechanisms (names,codes or shortnames)
along with the other required paramaters. It will return a vector of UIDs.
}
\examples{
\dontrun{
d<-d2Parser("myfile.csv",type="csv")
#Add a new column with organisation unit short names. 
d$mech_names<-remapMechs(d$attributeOptionCombo,mode_in="id",mode_out="code")
}
}
