% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remapOUs.R
\name{remapOUs}
\alias{remapOUs}
\title{Function which converts mechanism codes, names or shortnames to UIDs}
\usage{
remapOUs(ous_in, base.url, username, password, mode, ou)
}
\arguments{
\item{ous_in}{A vector of organisation unit identifiers (codes, names or shortNames)}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{mode}{Should be one of code, name or shortName. This is the class we are mapping from to UIDs.}

\item{ou}{UID of the Operating Unit/Country}
}
\value{
Returns a vector of organisation unit UIDs
}
\description{
remapOUs should be supplied a vector of organisation Units (names,codes or shortnames)
along with the other required paramaters. It will return a vector of UIDs.
}
\note{
remapOUs(foo,"https://www.datim.org","admin","district","code","Ab12345678")
will remap organisation units specified as codes to UIDs
}

