% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runValidation.R
\name{runValidation}
\alias{runValidation}
\title{Title}
\usage{
runValidation(
  d,
  datasets = NA,
  organisationUnit = NA,
  datastream = NA,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2session}{}
}
\description{
Title
}
