% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{saveCachedObject}
\alias{saveCachedObject}
\title{saveCachedObject(toCache,sig,wd)}
\usage{
saveCachedObject(toCache, sig, wd = paste0(getwd(), "/.R_Cache/"))
}
\arguments{
\item{toCache}{Object which should be cached.}

\item{sig}{MD5 hash of the objects URL reference}

\item{wd}{By default, the .R_Cache directory in the working directory.}
}
\value{
Nothing.
}
\description{
Internal utility function to save an object to the cache.
}

