% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataset.R
\name{selectDataset}
\alias{selectDataset}
\title{selectDataSet()}
\usage{
selectDataset()
}
\value{
Returns a data frame  of "dataElementName","categoryOptionComboName","dataElement","categoryOptionCombo"
of invalid data elements which are present the the data
}
\description{
Utility function to produce a vector of data set UIDs from user input
}

