% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftSIMSData.R
\name{shiftSIMSData}
\alias{shiftSIMSData}
\title{Title}
\usage{
shiftSIMSData(d)
}
\arguments{
\item{d}{datimvalidation object}
}
\value{
datimvalidation object with with dates deconflicted in 
d$data$import.
}
\description{
Title
}
