% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2Parser.R
\name{sims2Parser}
\alias{sims2Parser}
\title{Title}
\usage{
sims2Parser(...)
}
\arguments{
\item{...}{Same paramaters as d2Parser, except the mode is set to "SIMS"}
}
\value{
Returns a data frame of at least "dataElement","period","orgUnit","categoryOptionCombo","attributeOptionCombo","value"
}
\description{
Title
}
\seealso{
d2Parser
}
