% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validateData.R
\name{validateData}
\alias{validateData}
\title{Function which checks the of a DATIM data payload agaist define validation rules}
\usage{
validateData(data, base.url, username, password,
  return_violations_only = TRUE, parallel = TRUE, outputToFile)
}
\arguments{
\item{data}{d2Parser data frame}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{return_violations_only}{Paramater to return only violations or all validation rule evalualtions.}

\item{parallel}{Should the rules be evaluated in parallel.}

\item{outputToFile}{Do not return an object, but rather output to a file directly}

\item{return_violations_only}{Only return violations}
}
\value{
Returns a data frame with validation rule results.
}
\description{
validateData should be supplied a d2Parser compliant data frame.
The data frame is checked dynamically against validation rules defined in the DATIM server.
}

