% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementMechanismValidity.R
\name{validateDEs_ACOCs}
\alias{validateDEs_ACOCs}
\title{Title Given a data frame of data elements and attribute option combinations
identify the invalid combinations.}
\usage{
validateDEs_ACOCs(des_acocs, des_acocs_map)
}
\arguments{
\item{des_acocs}{A data frame consisting of unique data element
attribute option combinations.}

\item{des_acocs_map}{A list of all possible valid Data elements
and attribute option combinations.}
}
\description{
Title Given a data frame of data elements and attribute option combinations
identify the invalid combinations.
}
