% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{validateOrgunitDataElements}
\alias{validateOrgunitDataElements}
\title{Title}
\usage{
validateOrgunitDataElements(orgunit_data_elements, de_map)
}
\arguments{
\item{de_map}{}
}
\description{
Title
}
