% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2Parser.R
\name{checkCodingScheme}
\alias{checkCodingScheme}
\title{Utility function to check that the supplied coding scheme is correct}
\usage{
checkCodingScheme(data, d2session = d2_default_session)
}
\arguments{
\item{data}{A parsed DHIS2 data payload from d2parser}

\item{d2session}{datimutils d2session object}
}
\value{
Warnings are issued if the coding scheme is not 
congruent with what has been supplied as a paramater.
}
\description{
checkCodingScheme will ensure that all indentifiers 
after parsing are valid UIDs,
or in the case of periods, valid periods.
}
