% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementDisaggValidity.R
\name{checkDataElementDisaggValidity}
\alias{checkDataElementDisaggValidity}
\title{checkDataElementDisaggValidity(data,dataset)}
\usage{
checkDataElementDisaggValidity(data, datasets = NA,
  return_violations = TRUE)
}
\arguments{
\item{data}{D2 compliant data frame object}

\item{datasets}{Should be a character vector of data set UIDs. 
Alternatively, if left missing, user will be promted.}

\item{return_violations}{Boolean to return violations only. 
 @return Returns a data frame  of "dataElementName","categoryOptionComboName",
"dataElement","categoryOptionCombo"
of invalid data elements which are present the the data}
}
\description{
Utility function to produce a data frame of
invalid data elements based on current
DATIM form specification
}
