% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{checkDataElementOrgunitValidity}
\alias{checkDataElementOrgunitValidity}
\title{checkDataElementOrgunitValidity(data,organisationUnit,datasets, d2session)}
\usage{
checkDataElementOrgunitValidity(
  d,
  datasets,
  return_violations = TRUE,
  d2session = d2_default_session
)
}
\arguments{
\item{datasets}{Should be a character vector of data set UIDs.}

\item{return_violations}{Return the invalid data if TRUE}

\item{d2session}{datimutils d2session object}

\item{data}{D2 compliant data frame}
}
\value{
Returns subset of data which contains
  invalid data element / organisation unit associations. If no violations are found, a boolean
  TRUE value is returned.
}
\description{
Returns a data frame invalid data elements which exist in the data
but which do not have a valid organistion unit / dataset association.
}
\examples{
\dontrun{
     d<-d2Parser("myfile.csv",type="csv")
     ds<-getCurrentMERDataSets(type="RESULTS")
     checkDataElementOrgunitValidity(data=d,datasets=ds)
}

}
