% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{checkDataElementOrgunitValidity}
\alias{checkDataElementOrgunitValidity}
\title{checkDataElementOrgunitValidity(data,organisationUnit,datasets)}
\usage{
checkDataElementOrgunitValidity(
  data = NA,
  organisationUnit = NA,
  datasets = NA,
  return_violations = TRUE,
  creds
)
}
\arguments{
\item{data}{D2 compliant data frame}

\item{organisationUnit}{Should be the UID of the organisation unit ancestor, 
typically the operating unit.}

\item{datasets}{Should be a character vector of data set UIDs. 
Alternatively, if left missing, user will be promted.}

\item{return_violations}{Return the invalid data if TRUE}
}
\value{
Returns subset of data which contains
  invalid data element / organisation unit associations. If no violations are found, a boolean
  TRUE value is returned.
}
\description{
Returns a data frame invalid data elements which exist in the data 
but which do not have a valid organistion unit / dataset association.
}
\examples{
\dontrun{
     d<-d2Parser("myfile.csv",type="csv")
     ds<-getCurrentMERDataSets(type="RESULTS")
     checkDataElementOrgunitValidity(data=d,datasets=ds)
}
}
