% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataElementOrgunitValidity.R
\name{checkDataElementOrgunitValidity}
\alias{checkDataElementOrgunitValidity}
\title{checkDataElementOrgunitValidity(data,organisationUnit,datasets, d2session)}
\usage{
checkDataElementOrgunitValidity(
  d,
  datasets,
  return_violations = TRUE,
  d2session = d2_default_session
)
}
\arguments{
\item{d}{datimvalidation object obtained from d2Parser.}

\item{datasets}{Should be a character vector of data set UIDs.}

\item{d2session}{datimutils d2session object}
}
\value{
datimvalidation object obtained from d2Parser which have 
been checked for invalid data element/ organisation unit combinations
}
\description{
Returns a data frame invalid data elements which exist in the data
but which do not have a valid organistion unit / dataset association.
}
\examples{
\dontrun{
     d<-d2Parser("myfile.csv",type="csv")
     ds<-getCurrentDataSets(type="RESULTS")
     d <- checkDataElementOrgunitValidity(d,datasets=ds)
}

}
