% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMechanismValidity.R
\name{checkMechanismValidity}
\alias{checkMechanismValidity}
\title{checkMechanismPeriodValidity(data,organisationUnit)}
\usage{
checkMechanismValidity(
  data,
  organisationUnit = NA,
  return_violations = TRUE,
  d2session = d2_default_session
)
}
\arguments{
\item{data}{A data frame which has been parsed by either d2Parser or sims2Parser}

\item{organisationUnit}{UID of the operating unit.}

\item{return_violations}{Should the function return a list of violations only.}

\item{d2session}{datimutils d2session object}
}
\value{
Returns a data frame of invalid period-mechanism combinations. A warning will 
also be issued. 
Returns TRUE if there are no violations.
}
\description{
This function will return an object of invalid mechanisms and periods. 
All data which is reported 
must have a period within the valid start and end dates of the
attribute option combination to which it is assigned. 
The mechanism must also be associated with the operating unit. 
If either of these two conditions 
are not met, the data will be flagged as being invalid.
}
\examples{
\dontrun{
     d<-d2Parser("myfile.csv",type="csv")
     ds<-getCurrentMERDataSets(type="RESULTS")
     checkMechanismValidity(data=d, organisationUnit = "f5RoebaDLMx")
}

}
