% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNarrativeLength.R
\name{checkNarrativeLength}
\alias{checkNarrativeLength}
\title{Title Check that the length of narrative data does not exceed 50k characters.}
\usage{
checkNarrativeLength(d)
}
\arguments{
\item{d}{datimvalidation object obtained from d2Parser.}
}
\value{
datimvalidation object obtained from d2Parser
checked for long narratives.
}
\description{
Title Check that the length of narrative data does not exceed 50k characters.
}
