% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNegativeValues.R
\name{checkNegativeValues}
\alias{checkNegativeValues}
\title{checkNegativeValues(data)}
\usage{
checkNegativeValues(data, d2session = d2_default_session)
}
\arguments{
\item{data}{D2 compliant data frame object}

\item{d2session}{datimutils d2session object}
}
\value{
A filtered D2 compliant data frame of invalid values, otherwise, 
returns TRUE if there are no violations. A warning will also be issued if 
there are any negative values found in non-dedupe mechanisms.
}
\description{
In general, data values which are negative will not be imported.
}
\examples{
\dontrun{
 d<-d2Parser("myfile.csv",type="csv")
 checkNegativeValues(data=d)
}

}
