% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNegativeValues.R
\name{checkNegativeValues}
\alias{checkNegativeValues}
\title{checkNegativeValues(data)}
\usage{
checkNegativeValues(d)
}
\arguments{
\item{data}{D2 compliant data frame object}
}
\value{
A filtered D2 compliant data frame of invalid values, otherwise, returns NA.
}
\description{
In general, we should never accept data values
for numeric data elements which are negative, unless they are 
deduplication data, whose import is currently not supported.
}
