% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOperatingUnit.R
\name{checkOperatingUnit}
\alias{checkOperatingUnit}
\title{Function which checks that the organisation unit provided is valid}
\usage{
checkOperatingUnit(organisationUnit = NA)
}
\arguments{
\item{organisationUnit}{Should be a UID of an operating unit. Defaults to the 
user organisation unit if not supplied}
}
\value{
Returns a vector of organisation unit UIDs
}
\description{
Validates that the supplied UID is a valid operating unit UID
}
\note{
remapOUs(foo,"https://www.datim.org","admin","district","code","id","Ab12345678")
will remap organisation units specified as codes to UIDs
}
