% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{checkOptionSetCompliance}
\alias{checkOptionSetCompliance}
\title{checkOptionSetCompliance()}
\usage{
checkOptionSetCompliance(d, creds)
}
\arguments{
\item{d}{D2Parsed data frame}

\item{creds}{DHISLogin object}
}
\value{
Returns a data frame of invalid values validated against their option set.
}
\description{
Internal function for validation of data which have option sets
}
\examples{
 \dontrun{
  d<-d2Parser("myfile.csv",type="csv")
  checkOptionSetCompliance(d)
}
}
