% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOrgunitsInHierarchy.R
\name{checkOrgunitsInHierarchy}
\alias{checkOrgunitsInHierarchy}
\title{Title}
\usage{
checkOrgunitsInHierarchy(
  d,
  userOrgUnit,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{datimvalidation object obtained from d2Parser.}

\item{userOrgUnit}{UID of the operating unit which the user belongs to, or wishes to assume}

\item{d2session}{datimutils d2session object}
}
\value{
datimvalidation object
}
\description{
Title
}
