% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{checkValueTypeCompliance}
\alias{checkValueTypeCompliance}
\title{checkValueTypeCompliance(d)}
\usage{
checkValueTypeCompliance(d)
}
\arguments{
\item{d}{A D2Parsed data frame}
}
\value{
Returns a data frame of invalid data only
}
\description{
Utility function of extraction of data element ids, codes, shortName and names
}
