% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{checkValueTypeCompliance}
\alias{checkValueTypeCompliance}
\title{checkValueTypeCompliancep(d,base.url,username,passowrd)}
\usage{
checkValueTypeCompliance(d, base.url, username, password)
}
\arguments{
\item{d}{Parsed data frame}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a parsed data frame plus the valueType
}
\description{
Utility function of extraction of data element ids, codes, shortName and names
}

