% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{checkValueTypeCompliance}
\alias{checkValueTypeCompliance}
\title{checkValueTypeCompliance(d)}
\usage{
checkValueTypeCompliance(d, d2session = d2_default_session)
}
\arguments{
\item{d}{D2 Parsed data frame}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame of invalid data only
}
\description{
This function will check data to ensure that its value matches
a given pattern based on its value type. As an example, if the data element
which is being validated should only accept positive integers, any values
which are not positive integers will be flagged. For values which are composed
of option sets, any values which do not correspond so one of the option set 
value codes, will also be flagged. Additionally ,any values which are zero
for which the data element is not zero significant will also be flagged.
}
\examples{
\dontrun{
d<-d2Parser("myfile.csv",type="csv")
checkValueTypeCompliance(d)
}

}
