% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{clearCache}
\alias{clearCache}
\title{clearCache(wd)}
\usage{
clearCache(cache_dir = paste0(getwd(), "/.R_Cache/"), force = NA)
}
\arguments{
\item{cache_dir}{Cache directory path. By default, working directory + ".R_Cache"}

\item{force}{Boolean (default FALSE) to clear cache without asking the user}
}
\value{
Nothing.
}
\description{
Internal utility function to clear the object cache.
}
