% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2_api_get.R
\name{d2_api_get}
\alias{d2_api_get}
\title{Execute and return a DATIM API query.}
\usage{
d2_api_get(path, d2session, timeout = 60)
}
\arguments{
\item{path}{Should begin with api/ and contain the query}

\item{timeout}{how long should a reponse be waited for}

\item{d2_session}{the d2Session object, default is "d2_default_session",
it will be made upon logining in to datim with loginToDATIM}

\item{retry}{number of times to try in case of failure,
default will not try again}

\item{api_version}{defaults to current but can pass in version number}

\item{verbose}{return raw content with data}

\item{quiet}{Echo the URL which is called to the console.}
}
\value{
Result of DATIM API query returned as named list.

Copied almost verbatim from datimutils
}
\description{
Gets and flattens DATIM API query as dataframe.
}
