% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSecrets.R
\name{DHISLogin}
\alias{DHISLogin}
\title{DHISLogin(config_path)}
\usage{
DHISLogin(dhis_config)
}
\arguments{
\item{dhis_config}{List of DHIS2 credentials}
}
\value{
TRUE if you are able to login to the server.
}
\description{
DHISLogin(config_path)
}
