% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluateValidation.R
\name{evaluateValidation}
\alias{evaluateValidation}
\title{Utility function for evaluating validation rules}
\usage{
evaluateValidation(combis, values, vr, violationsOnly = TRUE, outputFile)
}
\arguments{
\item{combis}{Data elements and category option combinations}

\item{values}{Values}

\item{vr}{Validation rule object}

\item{violationsOnly}{Only return validation rule violations}

\item{Output}{to a file directly instead of returning an object}
}
\value{
Returns a data frame of validation rule evaluations, alternatively a file
}
\description{
evaluationValidation will accept a vector of combis (data elements.category optioncombinations) and values
and evaluate it against the supplied validation rules
}

