% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{getCachedObject}
\alias{getCachedObject}
\title{Get Cached Object}
\usage{
getCachedObject(sig, wd = paste0(getwd(), "/.R_Cache/"))
}
\arguments{
\item{sig}{MD5 hash of the objects URL reference}

\item{wd}{By default, the .R_Cache directory in the working directory.}
}
\value{
Returns a cached data object.
}
\description{
Internal utility function to retrieve a cached object.
}
