% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCategoryOptionCombosMap.R
\name{getCategoryOptionCombosMap}
\alias{getCategoryOptionCombosMap}
\title{getCategoryOptionCombosMap(base.url,username,passowrd)}
\usage{
getCategoryOptionCombosMap(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a data frame  of code,name,id and shortName of all categoryOptionCombos
}
\description{
Utility function to produce a map of category option combos
}

