% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCategoryOptionCombosMap.R
\name{getCategoryOptionCombosMap}
\alias{getCategoryOptionCombosMap}
\title{getCategoryOptionCombosMap()}
\usage{
getCategoryOptionCombosMap(d2session = d2_default_session)
}
\arguments{
\item{d2session}{datimutils d2session object}
}
\value{
Returns a data frame  of code,name,id and shortName of all categoryOptionCombos
}
\description{
Utility function to produce a map of category option combos
}
\examples{
\dontrun{
coc_map<-getCategoryOptionCombosMap()
}
}
