% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrossWalkDuplicates}
\alias{getCrossWalkDuplicates}
\title{Get Crosswalk Duplicates}
\usage{
getCrossWalkDuplicates(
  d,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{D2 Parsed data frame}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame of all crosswalk mechanism values,
 along with default de-deudplication adjustments.
}
\description{
Utility function which returns the crosswalk mechanism UID
}
\examples{
\dontrun{
    cw_dups <- getCrossWalkDuplicates(d)
 }


}
