% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrossWalkDuplicates}
\alias{getCrossWalkDuplicates}
\title{getCrossWalkDuplicates(d)}
\usage{
getCrossWalkDuplicates(d)
}
\arguments{
\item{d}{A d2 parsed data frame of data.}
}
\value{
Returns a data frame of all crosswalk mechanism values,
 along with default de-deudplication adjustments.
 @examples 
  \dontrun{
    cw_dups<-getCrossWalkDuplicates(d)
 }
}
\description{
Utility function which returns the crosswalk mechanism UID
}
