% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMap}
\alias{getCrosswalkMap}
\title{Get Crosswalk Map}
\usage{
getCrosswalkMap(d2session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2session}{DHIS2 Session id}
}
\value{
Returns a map of DSD/TA data element UIDs.
}
\description{
Utility function which returns a map of
DSD / TA data elements UIDs
}
\examples{
\dontrun{
    cw_map <- getCrosswalkMap()
 }

}
