% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMap}
\alias{getCrosswalkMap}
\title{getCrosswalkMap()}
\usage{
getCrosswalkMap()
}
\value{
Returns a map of DSD/TA data element UIDs.
}
\description{
Utility function which returns a map of DSD / TA data elements UIDs
}
\examples{
\dontrun{
    cw_map<-getCrosswalkMap()
 }

}
