% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMechanism}
\alias{getCrosswalkMechanism}
\title{getCrosswalkMechanism(d2session = d2_default_session)}
\usage{
getCrosswalkMechanism(d2session = d2_default_session)
}
\value{
Returns a UID of the crosswalk de-duplication mechanism
}
\description{
Utility function which returns the crosswalk mechanism UID
}
\examples{
\dontrun{
    cw_uid<-getCrosswalkMechanism()
 }


}
