% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getCrosswalkMechanism}
\alias{getCrosswalkMechanism}
\title{getCrosswalkMechanism(base.url,username,password)}
\usage{
getCrosswalkMechanism(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a UID of the crosswalk de-duplication mechanism
}
\description{
Utility function which returns the crosswalk mechanism UID
}

