% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getCurrentMERDataSets}
\alias{getCurrentMERDataSets}
\title{Get Current MER Data Sets}
\usage{
getCurrentMERDataSets(
  type = "RESULTS",
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{type}{Should be either RESULTS or TARGETS}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a list of dataset UIDs of the given type
}
\description{
Get Current MER Data Sets
}
\examples{
  \dontrun{
 ds <- getCurrentMERDataSets(type="RESULTS", d2session = d2session)
 }

}
