% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getCurrentMERDataSets}
\alias{getCurrentMERDataSets}
\title{getCurrentMERDataSets(type)}
\usage{
getCurrentMERDataSets(type = "RESULTS")
}
\arguments{
\item{type}{Should be either RESULTS or TARGETS}
}
\value{
Returns a list of dataset UIDs of the given type.
}
\description{
Get a list of MER datasets which are currently open for data entry.
}
\examples{
 \dontrun{
 ds<-getCurrentMERDataSets(type="RESULTS")
 }
}
