% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getCurrentMERDataSets}
\alias{getCurrentMERDataSets}
\title{getCurrentMERDataSets(type)}
\usage{
getCurrentMERDataSets(type = "RESULTS", d2session = d2_default_session)
}
\arguments{
\item{type}{Should be either RESULTS or TARGETS}

\item{d2session}{datimutils login session object}
}
\value{
Returns a list of dataset UIDs of the given type
}
\description{
getCurrentMERDataSets(type)
}
\examples{
  \dontrun{
 ds<-getCurrentMERDataSets(type="RESULTS", d2session = d2session)
 }

}
