% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataElementMap.R
\name{getDataElementMap}
\alias{getDataElementMap}
\title{getDataElementMap()}
\usage{
getDataElementMap(creds)
}
\arguments{
\item{creds}{DHISLogin object.}
}
\value{
Returns a data frame  of id,code,shortName and name
}
\description{
Utility function of extraction of data element ids, codes, shortName and names
}
\examples{
\dontrun{
de_map<-getDataElementMap(creds = my_creds)
}
}
