% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidDatasetMembers.R
\name{getDataElementsOrgunits}
\alias{getDataElementsOrgunits}
\title{getInvalidDatasetMembers(data,base.url,username,password,organisationUnit,dataset)}
\usage{
getDataElementsOrgunits(base.url, username, password, organisationUnit, dataset)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{organisationUnit}{Should be the UID of the organisation unit ancestor.}

\item{dataset}{Should be a description of the dataset group, such as "MER Results", "MER Targets" or "SIMS"}
}
\value{
Returns a data frame of "data" which was submitted along for a reason of why the data is considered to be invalid.
}
\description{
Returns a data frame of invalid data element /organisation unit combinations
based on the current data set assignments
}

