% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getDataSets}
\alias{getDataSets}
\title{getDataSets(creds)}
\usage{
getDataSets(creds)
}
\value{
Returns a data frame  of name, id and formtype of all datasets.
}
\description{
Utility function to produce a data frame of
 datasets. 
 @param creds DHISLogin object.
}
\examples{
 \dontrun{
 ds<-getDataSets()
 }

}
