% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSets.R
\name{getDataSets}
\alias{getDataSets}
\title{Get Data Sets}
\usage{
getDataSets(d2session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame  of name, id and formtype of all datasets.
}
\description{
Utility function to produce a data frame of datasets.
}
\examples{
 \dontrun{
 ds <- getDataSets()
 }

}
