% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getExactDuplicates}
\alias{getExactDuplicates}
\title{getExactDuplicates(data)}
\usage{
getExactDuplicates(d)
}
\arguments{
\item{d}{D2 Parsed data frame}
}
\value{
Returns a data frame of d2 data, with exact duplicates.
  A warning will be issued if duplicates are detected.
}
\description{
Returns a listing of exact duplicates by dataElement,
period,orgunit,categoryOptionCombo and attributeOptionCombo
}
\examples{
\dontrun{
 d<-d2Parser("myfile.csv",type="csv",header=TRUE)
 dups<-getExactDuplicates(d)
 dups
 }
}
