% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidDataElements.R
\name{getInvalidDataElements}
\alias{getInvalidDataElements}
\title{getInvalidDataElements(data,base.url,username,password,dataset)}
\usage{
getInvalidDataElements(data, base.url, username, password, datasets = NA)
}
\arguments{
\item{data}{D2 compliant data frame object}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{datasets}{Should be a character vector of data set UIDs. Alternatively, if left missing, user will be promted.}
}
\value{
Returns a data frame  of "dataElementName","categoryOptionComboName","dataElement","categoryOptionCombo"
of invalid data elements which are present the the data
}
\description{
Utility function to produce a data frame of invalid data elements based on current
DATIM form specification
}

