% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvalidMechanisms.R
\name{getInvalidMechanisms}
\alias{getInvalidMechanisms}
\title{Function which checks the validity of mechanisms}
\usage{
getInvalidMechanisms(data, organisationUnit = NA)
}
\arguments{
\item{data}{d2Parser data frame}

\item{organisationUnit}{UID of the operating unit.Defaults to the user organisation unit if left blank}
}
\value{
Returns a vector of non-valid mechanisms
}
\description{
checkValidMechanisms should be supplied a d2Parser compliant data frame along with the operating unit UID
along with the other required paramaters. It will return a vector of non-valid mechanism UIDs which were part of the data
}
\note{
getValidMechanisms(foo,"https://www.datim.org","admin","district","Ab12345678")
will remap organisation units specified as codes to UIDs
}
