% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechanismsMap.R
\name{getMechanismsMap}
\alias{getMechanismsMap}
\title{getMechanismsMap()}
\usage{
getMechanismsMap(organisationUnit = NA, endDate = "2016-09-29")
}
\arguments{
\item{organisationUnit}{Operating unit in question specified as a UID. If not specified, assume to be Global.}

\item{endDate}{Date which marks the mechanism as valid. Must be specified in a valid R date syntax}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}

