% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechanismsMap.R
\name{getMechanismsMap}
\alias{getMechanismsMap}
\title{getMechanismsMap()}
\usage{
getMechanismsMap(organisationUnit = NA)
}
\arguments{
\item{organisationUnit}{Operating unit in question specified as a UID. If not specified, assume to be Global.}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest) along with start and endDates of the mechanisms
}
\description{
Utility function to produce a map of valid mechanisms
}

