% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechanismsMap.R
\name{getMechanismsMap}
\alias{getMechanismsMap}
\title{Get Mechanisms Map}
\usage{
getMechanismsMap(
  organisationUnit = NA,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{organisationUnit}{Organisation unit ancestor.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo
(which is the UID of interest) along with start and endDates of the mechanisms
}
\description{
Utility function to produce a map of valid mechanisms. If no
organisation unit is provided, assumed to be global.
}
