% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMechanismsMap.R
\name{getMechanismsMap}
\alias{getMechanismsMap}
\title{getMechanismsMap(base.url,username,passowrd)}
\usage{
getMechanismsMap(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}

