% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{getOptionSetMap}
\alias{getOptionSetMap}
\title{getOptionSetMap()}
\usage{
getOptionSetMap()
}
\value{
Returns a list of option sets values per option set
}
\description{
Utility function to fetch a list of optionsets and possible values within them
}
