% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance.R
\name{getOptionSetMap}
\alias{getOptionSetMap}
\title{getOptionSetMap()}
\usage{
getOptionSetMap(d2session = d2_default_session)
}
\arguments{
\item{d2session}{datimutils d2session object}
}
\value{
Returns a data frame structure composed of option set names, UIDs, and a data frame for
each of these with individual options present in the option set.
}
\description{
Utility function to fetch a list of optionsets and possible values within them
}
\examples{
 \dontrun{
  op_set_map<-getOptionSetMap()
}
}
