% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganisationUnitMap.R
\name{getOrganisationUnitMap}
\alias{getOrganisationUnitMap}
\title{getOrganisationUnitMap(operatingUnit)}
\usage{
getOrganisationUnitMap(organisationUnit = NA)
}
\arguments{
\item{organisationUnit}{of question specified as a UID}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}

