% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganisationUnitMap.R
\name{getOrganisationUnitMap}
\alias{getOrganisationUnitMap}
\title{getOrganisationUnitMap(operatingUnit)}
\usage{
getOrganisationUnitMap(organisationUnit = NA, creds)
}
\arguments{
\item{organisationUnit}{of question specified as a UID}

\item{creds}{DHISLogin object}
}
\value{
Returns a data frame  of name,code,id, and categoryOptionCombo (which is the UID of interest)
}
\description{
Utility function to produce a map of valid mechanisms
}
\examples{
\dontrun{
    ou_map<-getOrganisationUnitMap("KKFzPM8LoXs") #For a specific operating unit
    ou_map_mine<-getOrganisationUnitMap() #Based on your login credentials
}
}
