% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriodInfo.R
\name{getPeriodInfo}
\alias{getPeriodInfo}
\title{getPeriodInfo(ISO)}
\usage{
getPeriodInfo(ISO = NA, d2session = d2_default_session)
}
\arguments{
\item{ISO}{Formatted string such as 2016Q1. Refer to DHIS2 documentation for details.}

\item{d2session}{datimutils d2session object}
}
\value{
Returns a data frame of  ISO, startDate,endDate,and periodtype
}
\description{
Get information like start date, end date and period type from an ISO period string
}
