% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodUtils.R
\name{getPeriodType}
\alias{getPeriodType}
\title{getPeriodType(iso)}
\usage{
getPeriodType(iso)
}
\arguments{
\item{ISO}{String which identifies the period, such as 2016Q1 or 2016Q2}
}
\value{
Returns a a character of the period, like "Monthly"
}
\description{
Utility function return a DHIS2 period type from an ISO string
}

