% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDuplication.R
\name{getPureDuplicates}
\alias{getPureDuplicates}
\title{getPureDuplicates(data)}
\usage{
getPureDuplicates(d)
}
\arguments{
\item{d}{A d2 parsed data frame}
}
\value{
Returns a data frame of d2 data, with pure duplicates.
}
\description{
Returns a listing of pure duplicates by dataElement,period,orgunit,categoryOptionCombo. 
For the purposes of de-duplication, these are considered to be pure duplicates (only the attributeOptionCombo differ)
}

