% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidDataElements.R
\name{getValidDataElements}
\alias{getValidDataElements}
\title{getInvalidDataElements(data,base.url,username,password,dataset)}
\usage{
getValidDataElements(base.url, username, password, dataset)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{dataset}{Should be a description of the dataset group, such as "MER Results", "MER Targets" or "SIMS"}
}
\value{
Returns a data frame  of "dataElementName","categoryOptionComboName","dataElement","categoryOptionCombo"
of invalid data elements which are present the the data
}
\description{
Utility function to produce a data frame of valid data elements based on current
DATIM form specification
}

