% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidDataElements.R
\name{getValidDataElements}
\alias{getValidDataElements}
\title{getInvalidDataElements(datasets)}
\usage{
getValidDataElements(datasets = NA, d2session = d2_default_session)
}
\arguments{
\item{datasets}{Should be a character vector of data set UIDs. Alternatively, if left missing, user will be promted.}

\item{d2session}{datimutils login session object}
}
\value{
Returns a data frame  of "dataSet","dataElementName","shortname","code","dataelementuid","categoryOptionComboName"
}
\description{
Utility function to produce a data frame of valid data elements based on current
DATIM form specification
}
