% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidDataElements.R
\name{getValidDataElements}
\alias{getValidDataElements}
\title{Get Invalid Data Elements}
\usage{
getValidDataElements(
  datasets = NA,
  d2session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{datasets}{Should be a character vector of data set UIDs.
Alternatively, if left missing, user will be promted to choose from a list.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame  of "dataSet", "dataElementName",
"shortname", "code", "dataelementuid", "categoryOptionComboName"
}
\description{
Utility function to produce a data frame of valid data elements based on current
DATIM form specification
}
