% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getValidOperatingUnits.R
\name{getValidOperatingUnits}
\alias{getValidOperatingUnits}
\title{Utility function which provides a list of valid operating units}
\usage{
getValidOperatingUnits(base.url, username, password)
}
\arguments{
\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}
}
\value{
Returns a data frame of operating units and their UIDs
}
\description{
Utility function which provides a list of valid operating units
}
\note{
getValidOperatingUnits("https://www.datim.org","admin","district")
will remap organisation units specified as codes to UIDs
}

