% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidOperatingUnits.R
\name{getValidOperatingUnits}
\alias{getValidOperatingUnits}
\title{Utility function which provides a list of valid operating units}
\usage{
getValidOperatingUnits(d2session = d2_default_session)
}
\arguments{
\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame of operating units and their UIDs
}
\description{
Utility function which provides a list of valid operating units
}
