% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSecrets.R
\name{LoadConfigFile}
\alias{LoadConfigFile}
\title{LoadConfig(config_path)}
\usage{
LoadConfigFile(config_path = NA)
}
\arguments{
\item{config_path}{Path to the DHIS2 credentials file}
}
\value{
A list of baseurl, username and password
}
\description{
Loads a JSON configuration file to access a DHIS2 instance
}
