% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSecrets.R
\name{loadSecrets}
\alias{loadSecrets}
\title{loadSecrets(secrets)}
\usage{
loadSecrets(config_path = NA)
}
\arguments{
\item{config_path}{Location of the configuration file to a DHIS2 server.}
}
\value{
Returns a boolean value indicating the user was able ot successfully 
login to the system
}
\description{
Load a DHIS secrets file which should be a JSON file as below
{
dhis": {
 "baseurl": "http://localhost:8080",
 "username": "admin",
 "password": "district"
 }
}
}
\examples{
 \dontrun{
 loadSecrets("/home/littlebobbytables/creds.json")
 }
}
