% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{prepDataForValidation}
\alias{prepDataForValidation}
\title{Internal function which prepares a parsed data frame for validation}
\usage{
prepDataForValidation(data)
}
\arguments{
\item{data}{D2 compliant data frame object}
}
\value{
Modifed data object with combis and totals appended.
}
\description{
Internal function which prepares a parsed data frame for validation
}
