% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{prepDataForValidation}
\alias{prepDataForValidation}
\title{Internal function which prepares a parsed data frame for validation}
\usage{
prepDataForValidation(d)
}
\arguments{
\item{d}{Data frame parsed by d2Parser}
}
\value{
Modifed data object with combis and totals appended
}
\description{
Internal function which prepares a parsed data frame for validation
}
