% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remapDEs.R
\name{remapDEs}
\alias{remapDEs}
\title{Function which converts data element codes, names or shortnames to UIDs}
\usage{
remapDEs(des_in, base.url, username, password, mode)
}
\arguments{
\item{des_in}{A vector of data element identifiers (codes, names or shortNames)}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{mode}{Should be one of code, name or shortName. This is the class we are mapping from to UIDs.}
}
\value{
Returns a vector of DATIM uids for each data element which needs to be remapped.
}
\description{
remapDes should be supplied a vector of data elements (names,codes or shortnames)
along with the other required paramaters. It will return a vector of UIDs
}
\note{
remapDes(foo,"https://www.datim.org","admin","district","code")
will remap data elements specified as codes to UIDs
}

