% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remapMechs.R
\name{remapMechs}
\alias{remapMechs}
\title{Function which converts mechanism codes, names or shortnames to UIDs}
\usage{
remapMechs(mechs_in, base.url, username, password, organisationUnit,
  mode_in = "code", mode_out = "id")
}
\arguments{
\item{mechs_in}{A vector of data element identifiers (codes, names or shortNames)}

\item{base.url}{Location of the server}

\item{username}{Server username}

\item{password}{Server password}

\item{mode_in}{Should be one of code, name or id. This is the class we are mapping from.}

\item{mode_out}{Should be one of code,name or id. This is the class we are mapping to.}

\item{ou}{The UID of the operating unit.}
}
\value{
Returns a vector of mechanism UIDs
}
\description{
remapMechs should be supplied a vector of data mechanisms (names,codes or shortnames)
along with the other required paramaters. It will return a vector of UIDs.
}
\note{
remapMechs(foo,"https://www.datim.org","admin","district","code","Ab12345678")
will remap mechanisms specified as codes to UIDs
}

