% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapOUs.R
\name{remapOUs}
\alias{remapOUs}
\title{Function which converts mechanism codes, names or shortnames to to another class of identifiers}
\usage{
remapOUs(
  ous_in,
  organisationUnit,
  mode_in = "code",
  mode_out = "id",
  d2session = d2_default_session
)
}
\arguments{
\item{ous_in}{A vector of organisation unit identifiers (codes, names, or
shortNames).}

\item{organisationUnit}{Organisation unit ancestor.}

\item{mode_in}{Should be one of code, name,shortName or id. This is the class
we are mapping from.}

\item{mode_out}{Should be one of code,name,shortName or id. This is the class
we are mapping to.}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a vector of organisation unit UIDs will remap organisation units specified as codes to UIDs
}
\description{
remapOUs should be supplied a vector of organisation Units (names,codes or shortnames)
along with the other required paramaters. It will return a vector of to another class of identifiers.
}
\examples{
\dontrun{
d<-d2Parser("myfile.csv",type="csv")
    d$ou_names<-remapOUs(d$organisationUnits,mode_in="id",mode_out="shortName")
}

}
