% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runValidation.R
\name{runValidation}
\alias{runValidation}
\title{Title}
\usage{
runValidation(d, d2session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d}{datimvalidation object}

\item{d2session}{datimutils session object}

\item{datasets}{list of dataset UIDs}

\item{organisationUnit}{users organisation unit or one which they wish to assume}

\item{datastream}{one of RESULTS, TARGETS, NARRATIVES, SIMS.}
}
\value{
datimvalidation object with all validations run.
}
\description{
Title
}
