% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{validateData}
\alias{validateData}
\title{Function which checks the of a DATIM data payload agaist define validation rules}
\usage{
validateData(data, organisationUnit = NA, return_violations_only = TRUE,
  parallel = TRUE)
}
\arguments{
\item{data}{d2Parser data frame}

\item{organisationUnit}{Organisation unit. Defaults to the user organisation unit if not supplied.}

\item{return_violations_only}{Paramater to return only violations or all validation rule evalualtions.}

\item{parallel}{Should the rules be evaluated in parallel.}
}
\value{
Returns a data frame with validation rule results.
}
\description{
validateData should be supplied a d2Parser compliant data frame.
The data frame is checked dynamically against validation rules defined in the DATIM server.
}

