% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{validateData}
\alias{validateData}
\title{Function which checks the of a DATIM data payload agaist define validation rules}
\usage{
validateData(
  data,
  organisationUnit = NA,
  return_violations_only = TRUE,
  parallel = FALSE,
  vr = NULL,
  d2session = d2_default_session
)
}
\arguments{
\item{data}{D2 compliant data frame object}

\item{organisationUnit}{Organisation unit ancestor.}

\item{return_violations_only}{Parameter to return only violations or all
validation rule evaluations.}

\item{parallel}{Should the rules be evaluated in parallel. Default is to not evaluate in parallel.}

\item{vr}{Validation rule object}

\item{d2session}{DHIS2 Session id}
}
\value{
Returns a data frame with validation rule results.
}
\description{
validateData should be supplied a d2Parser compliant data frame.
The data frame is checked dynamically against validation rules defined in the DATIM server.
}
\examples{
\dontrun{
  d<-d2Parser("myfile.csv",type="csv")
  vr_rules<-validateData(d)
  doMC::registerDoMC(cores=4)
  vr_rules<-validateData(d,parallel=TRUE)
  ds<-getCurrentMERDataSets(type="RESULTS")
  vr_rules<-validateData(d,parallel=TRUE,datasets=ds)
}
}
